requirements.txt
# Raspberry Pi 5 dependencies
# Install with: pip3 install -r requirements.txt
#
# lgpio replaces pigpio for Pi 5 (Pi 5 uses the RP1 GPIO chip; pigpio is incompatible).
# lgpio also works on Pi 4 — no conditional install needed.
#   sudo apt install python3-lgpio
#   pip3 install lgpio
#
# Pi 5 power note: the Pi 5 requires 5V/5A via USB-C PD. The VESC BEC must be
# rated for ≥5A continuous, or use a separate 5V/5A USB-C PD supply for the Pi.
#
# adafruit-blinka provides the board/busio abstractions for I2C:
#   sudo apt install python3-smbus i2c-tools

pyserial>=3.5
pyvesc>=1.0.2
adafruit-circuitpython-bno08x>=1.3.0
adafruit-blinka>=8.0.0
PyYAML>=6.0
lgpio>=0.2.2.0

# Web server
flask>=2.0
flask-cors>=3.0

# Testing (no hardware required)
pytest>=7.0
pytest-mock>=3.0
